Add Gemfile and puppet 4 checks
In anticipation of puppet 4, start trying to deal with puppet 4 things
that can be helpfully predicted by puppet lint plugins. Also fix lint
errors caught by the puppet-lint-absolute_classname-check and
puppet-lint-empty_string-check gem as well as arrow alignment errors
not caught before.

The has_variable? function in the layout.html.erb template will return
true whether $image is set to empty string or undef, so we change the
template to look for nil so that the new default of undef will work
correctly.

Change-Id: I6738cfef08e0cf2faeabed469b9eff5e3c0bbb5a